build: add Makefile
Makefile has a dependency on node and the smash package. Before
building, install node and run the following command:

    $ npm install smash

Then you can build elm-d3 using the following command:

    $ make

This will create a file called elm-d3.js which contains the compiled Elm
code as well as the elm-d3 runtime. It does not contain the Elm runtime,
nor does it include D3.

All the native code for elm-d3 uses smash to organize code.

Note that this makefile requires a version of the compile that supports
the `--src-dir` flag, which was introduced at evancz/elm@de52caca8ea6c6