runtime: fix sporadic error related to model update
The elm-d3 runtime sets up the model for a D3 custom element and then
later extends it in the render() callback. The update() callback uses
the added field, which is where the sporadic error occurs. The Elm
runtime occasionally decides to use the originally-setup model without
re-calling render(), leaving the model in a state that update() does not
expect.