widget: Rework selection API, introduce widgets
The previous implementation of bind with selections was awkward and
possibly not type-safe. To correct that, the latest API introduces
the Widget datatype:

  type Widget a b

  bind, (|=) : Selection a -> (a -> [b]) -> Widget a b
  chain', (|-) : Widget a b -> Selection b -> Widget a b

  embed : Widget a b -> Selection a

A Widget represent a `Selection a` whose data has been transformed to
some type `b`. The only way to perform operations on a Widget is using
the |- operator, which will apply the given selection to the widget.
Calling embed will freeze the widget and turn it back into a Selection
that can be combined with other Selections in the ususal way, e.g.,
method chaining (|.).