render: add static_ combinator
static_ is the equivalent of append except that it will append an
element only when its parent is created, and never after. Any selections
chained below the result will still be executed on every update.

Currently this is called static_ because of a bug in the Elm compiler.
Once that's resolved (already fixed in dev), it will be renamed
'static'.