render: Add nest operator
The nest operator is useful for composing views at the top level. It
will apply the first selection to the context to create a new context.
The second selection is then applied to the new context, with the new
context being the result of the entire expression.

It's not strictly necesary to implement nest nateively. It's actually
equivalent to the following:

  nest a b = a <.> (sequence update b)