static: fix index issues with static
When using static, operators that are nested below it do not receive the
proper index and associated datum. This is because static requires,
conceptually, a new data bind that will not preserve the index of the
context.

In order to preserve the index, static iterates over the elements of the
context selection, which also provides each element's index in the
context selection. The coditional element creation logic is run, and
then the selection continuation is run on the result, except the index
of the parent element is passed to the continuation. This will be
propagated through to all other operators below the static operation.