Operators return optional strings
For those operators that are setting some attribute-like thing–i.e.
attr, style, and property–there must be a way to completely clear the
value. In JavaScript, you accomplish this by using the null value. In
Elm, this corresponds to a Maybe type.

Uses of thse functions that used to look like this:

  attr "class" (\d i -> "cls-" ++ show i)

should be updated to look like this

  fun attr "class" (\d i -> "cls-" ++ show i)

Previous uses of fun, should be replaced with opt. Uses of num and str
are unaffected.