event: property must return a JsonValue
Handling checkboxes properly requires that you set the `checked`
attribute of the DOM element to something that's either truthy or falsy,
but not `null`. D3 interprets a `null` value as a signal to delete the
property, i.e.,

  delete element.check

which will break the checkbox. That means that an option type as a
return type for property is not an option. The alternative is either a
custom type, or just use JsonValue, which will be a valid value for any
property.

Go with that, then.