event: add TodoMVC example
To build, use the following commands:

  make clean && make
  elm --make --src-dir=src `./scripts/build-flags` `./scripts/include-css examples/todo.css` examples/Todo.elm

Note that the generated code will depend on D3 on the internet. If you
have a local copy you'd like to use, set the D3 environment variable
with its location. Note that you should prefix its location with
file://.

If you're on a mac, open it using this command:

  open build/examples/Todo.html

Note that the toggle button the upper-left hasn't been implemeneted
because that is the most pointless thing you could possibly add to a
TODO list you would actually want to use.