Fix for native API changes in elm-0.12.3
There were some native API changes in the 0.12.3 release of Elm that
broke elm-d3. Specifically, the module related to casting Elm values to
and from JavaScript values dropped most of its functions.

To maintain compatibility with previous runtimes, this commit introduces
the Native.D3.JavaScript module. It's a drop-in replacement for
Native.JavaScript that for both 0.12.3 runtimes, and those before it.
How far before is unknown, but most likely back to 0.10.

Thanks to @rylev for bringing this to my attention. Closes #6.